apps/server/requirements.txt
azure-functions
requests
openai
mysql-connector-python
PyJWT
bcrypt
azure-storage-blob